Remove JHawk from the project
We found that we couldn't map JHawk's metrics calculation results to
SSA's design pattern detection results in any reasonable way. This is
due to the fact that JHawk calculates its metrics from project SOURCES,
whereas SSA detects design patterns in project BINARIES.

While sources and binaries should be "similar", we found them to be
"too different" for our purposes in practice. Because of this, we
decided to move away from JHawk and on to ckjm-extended, which performs
its metrics calculation on project binaries rather than sources.